refactor: conversions file
