+++ rustfmt.toml
@@ -2,3 +2,9 @@ wrap_comments = true
 imports_granularity = "One"
 group_imports = "One"
 format_code_in_doc_comments = true
+match_arm_blocks = false
+blank_lines_lower_bound = 0
+blank_lines_upper_bound = 0
+condense_wildcard_suffixes = true
+error_on_unformatted = true
+error_on_line_overflow = true
