+++ src/allocationerror.rs
@@ -5,7 +5,6 @@ use {
         fmt::{Display, Formatter, Result as Format},
     },
 };
-
 /// Error type for APIs with fallible heap allocation
 #[derive(Debug)]
 pub enum AllocationError {
@@ -17,11 +16,9 @@ pub enum AllocationError {
         layout: Layout,
     },
 }
-
 impl Display for AllocationError {
     fn fmt(&self, f: &mut Formatter) -> Format {
         write!(f, "Allocation error: {:?}", self)
     }
 }
-
 impl Error for AllocationError {}
