+++ src/bytes.rs
@@ -2,43 +2,35 @@ use {
     super::SmallVec,
     bytes::{buf::UninitSlice, BufMut},
 };
-
 unsafe impl<const N: usize> BufMut for SmallVec<u8, N> {
     #[inline]
     fn remaining_mut(&self) -> usize {
         // A vector can never have more than isize::MAX bytes
         isize::MAX as usize - self.len()
     }
-
     #[inline]
     unsafe fn advance_mut(&mut self, cnt: usize) {
         let len = self.len();
         let remaining = self.capacity() - len;
-
         if remaining < cnt {
             panic!("advance out of bounds: the len is {remaining} but advancing by {cnt}");
         }
-
         // Addition will not overflow since the sum is at most the capacity.
         self.set_len(len + cnt);
     }
-
     #[inline]
     fn chunk_mut(&mut self) -> &mut UninitSlice {
         if self.capacity() == self.len() {
             self.reserve(64); // Grow the smallvec
         }
-
         let cap = self.capacity();
         let len = self.len();
-
         let ptr = self.as_mut_ptr();
         // SAFETY: Since `ptr` is valid for `cap` bytes, `ptr.add(len)` must be
         // valid for `cap - len` bytes. The subtraction will not underflow since
         // `len <= cap`.
         unsafe { UninitSlice::from_raw_parts_mut(ptr.add(len), cap - len) }
     }
-
     // Specialize these methods so they can skip checking `remaining_mut`
     // and `advance_mut`.
     #[inline]
@@ -48,7 +40,6 @@ unsafe impl<const N: usize> BufMut for SmallVec<u8, N> {
     {
         // In case the src isn't contiguous, reserve upfront.
         self.reserve(src.remaining());
-
         while src.has_remaining() {
             let s = src.chunk();
             let l = s.len();
@@ -56,12 +47,10 @@ unsafe impl<const N: usize> BufMut for SmallVec<u8, N> {
             src.advance(l);
         }
     }
-
     #[inline]
     fn put_slice(&mut self, src: &[u8]) {
         self.extend_from_slice(src);
     }
-
     #[inline]
     fn put_bytes(&mut self, val: u8, cnt: usize) {
         // If the addition overflows, then the `resize` will fail.
