+++ src/comparisons.rs
@@ -1,5 +1,4 @@
 use super::SmallVec;
-
 impl<T, U, const N: usize, const M: usize> PartialEq<SmallVec<U, M>> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -10,7 +9,6 @@ where
     }
 }
 impl<T, const N: usize> Eq for SmallVec<T, N> where T: Eq {}
-
 impl<T, U, const N: usize, const M: usize> PartialEq<[U; M]> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -20,7 +18,6 @@ where
         self[..] == other[..]
     }
 }
-
 impl<T, U, const N: usize, const M: usize> PartialEq<&[U; M]> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -30,7 +27,6 @@ where
         self[..] == other[..]
     }
 }
-
 impl<T, U, const N: usize> PartialEq<[U]> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -40,7 +36,6 @@ where
         self[..] == other[..]
     }
 }
-
 impl<T, U, const N: usize> PartialEq<&[U]> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -50,7 +45,6 @@ where
         self[..] == other[..]
     }
 }
-
 impl<T, U, const N: usize> PartialEq<&mut [U]> for SmallVec<T, N>
 where
     T: PartialEq<U>,
@@ -60,7 +54,6 @@ where
         self[..] == other[..]
     }
 }
-
 impl<T, const N: usize> PartialOrd for SmallVec<T, N>
 where
     T: PartialOrd,
@@ -70,7 +63,6 @@ where
         self.as_slice().partial_cmp(other.as_slice())
     }
 }
-
 impl<T, const N: usize> Ord for SmallVec<T, N>
 where
     T: Ord,
