+++ src/rawsmallvec.rs
@@ -6,7 +6,6 @@ use {
         ptr::{copy_nonoverlapping, NonNull},
     },
 };
-
 /// Either a stack array with `length <= N` or a heap array
 /// whose pointer and capacity are stored here.
 ///
@@ -17,10 +16,8 @@ pub union RawSmallVec<T, const N: usize> {
     pub inline: ManuallyDrop<MaybeUninit<[T; N]>>,
     pub heap: (NonNull<T>, usize),
 }
-
 impl<T, const N: usize> RawSmallVec<T, N> {
     pub const IS_ZST: bool = size_of::<T>() == 0;
-
     #[inline]
     pub const fn new() -> Self {
         Self::new_inline(MaybeUninit::uninit())
@@ -37,7 +34,6 @@ impl<T, const N: usize> RawSmallVec<T, N> {
             heap: (ptr, capacity),
         }
     }
-
     #[inline]
     pub const fn as_ptr_inline(&self) -> *const T {
         // SAFETY: it is safe because we aren't reading the value, just getting a
@@ -46,14 +42,12 @@ impl<T, const N: usize> RawSmallVec<T, N> {
         #[allow(unused_unsafe, reason = "requires unsafe in MSRV")]
         (unsafe { &raw const self.inline }).cast::<T>()
     }
-
     #[inline]
     pub const fn as_mut_ptr_inline(&mut self) -> *mut T {
         // SAFETY: same as above
         #[allow(unused_unsafe, reason = "requires unsafe in MSRV")]
         (unsafe { &raw mut self.inline }).cast::<T>()
     }
-
     /// # Safety
     ///
     /// The vector must be on the heap
@@ -61,7 +55,6 @@ impl<T, const N: usize> RawSmallVec<T, N> {
     pub const unsafe fn as_ptr_heap(&self) -> *const T {
         self.heap.0.as_ptr()
     }
-
     /// # Safety
     ///
     /// The vector must be on the heap
@@ -69,7 +62,6 @@ impl<T, const N: usize> RawSmallVec<T, N> {
     pub const unsafe fn as_mut_ptr_heap(&mut self) -> *mut T {
         self.heap.0.as_ptr()
     }
-
     /// # Safety
     ///
     /// `new_capacity` must be non zero, and greater or equal to the length.
@@ -83,7 +75,6 @@ impl<T, const N: usize> RawSmallVec<T, N> {
         debug_assert!(!Self::IS_ZST);
         debug_assert!(new_capacity > 0);
         debug_assert!(new_capacity >= len.value());
-
         let was_on_heap = len.on_heap();
         let ptr = if was_on_heap {
             self.as_mut_ptr_heap()
@@ -91,13 +82,11 @@ impl<T, const N: usize> RawSmallVec<T, N> {
             self.as_mut_ptr_inline()
         };
         let len = len.value();
-
         let new_layout =
             Layout::array::<T>(new_capacity).map_err(|_| AllocationError::CapacityOverflow)?;
         if new_layout.size() > isize::MAX as usize {
             return Err(AllocationError::CapacityOverflow);
         }
-
         let new_ptr = if len == 0 || !was_on_heap {
             // get a fresh allocation
             let new_ptr = alloc(new_layout) as *mut T; // `new_layout` has nonzero size.
@@ -107,12 +96,10 @@ impl<T, const N: usize> RawSmallVec<T, N> {
             new_ptr
         } else {
             // use realloc
-
             // this can't overflow since we already constructed an equivalent layout during
             // the previous allocation
             let old_layout =
                 Layout::from_size_align_unchecked(self.heap.1 * size_of::<T>(), align_of::<T>());
-
             // SAFETY: ptr was allocated with this allocator
             // old_layout is the same as the layout used to allocate the previous memory
             // block new_layout.size() is greater than zero
