+++ src/serde.rs
@@ -7,7 +7,6 @@ use {
         Deserialize, Deserializer, Serialize, Serializer,
     },
 };
-
 impl<T, const N: usize> Serialize for SmallVec<T, N>
 where
     T: Serialize,
@@ -20,7 +19,6 @@ where
         state.end()
     }
 }
-
 impl<'de, T, const N: usize> Deserialize<'de> for SmallVec<T, N>
 where
     T: Deserialize<'de>,
@@ -31,21 +29,17 @@ where
         })
     }
 }
-
 struct SmallVecVisitor<T, const N: usize> {
     phantom: PhantomData<T>,
 }
-
 impl<'de, T, const N: usize> Visitor<'de> for SmallVecVisitor<T, N>
 where
     T: Deserialize<'de>,
 {
     type Value = SmallVec<T, N>;
-
     fn expecting(&self, formatter: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         formatter.write_str("a sequence")
     }
-
     fn visit_seq<B>(self, mut seq: B) -> Result<Self::Value, B::Error>
     where
         B: SeqAccess<'de>,
@@ -54,11 +48,9 @@ where
         let len = seq.size_hint().unwrap_or(0);
         let mut values = SmallVec::new();
         values.try_reserve(len).map_err(B::Error::custom)?;
-
         while let Some(value) = seq.next_element()? {
             values.push(value);
         }
-
         Ok(values)
     }
 }
