+++ src/std.rs
@@ -1,7 +1,5 @@
 extern crate std;
-
 use {super::SmallVec, std::io};
-
 #[cfg(feature = "std")]
 #[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 impl<const N: usize> io::Write for SmallVec<u8, N> {
@@ -10,13 +8,11 @@ impl<const N: usize> io::Write for SmallVec<u8, N> {
         self.extend_from_slice(buf);
         Ok(buf.len())
     }
-
     #[inline]
     fn write_all(&mut self, buf: &[u8]) -> io::Result<()> {
         self.extend_from_slice(buf);
         Ok(())
     }
-
     #[inline]
     fn flush(&mut self) -> io::Result<()> {
         Ok(())
