+++ src/taggedlen.rs
@@ -1,5 +1,4 @@
 use core::marker::PhantomData;
-
 /// Vec guarantees that its length is always less than [`isize::MAX`] in
 /// *bytes*.
 ///
@@ -11,7 +10,6 @@ use core::marker::PhantomData;
 /// For a ZST, we never use the heap, so we just store the length directly.
 #[repr(transparent)]
 pub struct TaggedLen<T>(usize, PhantomData<T>);
-
 // Clone and Copy must be manually implemented because the generic interferes
 // with the derive attribute implementations.
 impl<T> Clone for TaggedLen<T> {
@@ -19,15 +17,12 @@ impl<T> Clone for TaggedLen<T> {
     fn clone(&self) -> Self {
         Self(self.0, PhantomData)
     }
-
     #[inline]
     fn clone_from(&mut self, source: &Self) {
         self.0 = source.0;
     }
 }
-
 impl<T> Copy for TaggedLen<T> {}
-
 impl<T> TaggedLen<T> {
     const IS_ZST: bool = size_of::<T>() == 0;
     #[inline]
@@ -40,7 +35,6 @@ impl<T> TaggedLen<T> {
             Self((len << 1) | on_heap as usize, PhantomData)
         }
     }
-
     #[inline]
     #[must_use]
     pub const fn on_heap(self) -> bool {
@@ -50,7 +44,6 @@ impl<T> TaggedLen<T> {
             (self.0 & 1_usize) == 1
         }
     }
-
     #[inline]
     pub const fn value(self) -> usize {
         if Self::IS_ZST {
